apps/home/admin.py
from django.contrib import admin
from django.db import models
from django_ckeditor_5.widgets import CKEditor5Widget
from modeltranslation.admin import TranslationAdmin
from .models import Home

@admin.register(Home)
class HomeAdmin(TranslationAdmin):
    list_display = ['title', 'text',]

    formfield_overrides = {
        models.TextField: {'widget': CKEditor5Widget(config_name='default')},
    }