requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.4.0
certifi==2025.11.12
distlib==0.4.0
Django==5.0.14
django-ckeditor-5==0.2.19
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
django-modeltranslation==0.19.16
environs==11.0.0
filelock==3.19.1
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
marshmallow==4.1.0
multidict==6.7.0
pillow==10.4.0
platformdirs==4.4.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.34.0
yarl==1.22.0
